refactor: move CSV loaders into their owning op modules
Each ``load_*_data`` function moves from ``perf_database.py`` into the
op module that owns the data it parses. ``perf_database.py`` keeps a
backward-compat re-export block so external callers of
``from aiconfigurator.sdk.perf_database import load_X_data`` keep
working unchanged.

Net effect: ``perf_database.py`` shrinks from 4,292 → 2,453 lines
(-1,839, -43%). The 25 loaders + 3 helpers (``_read_filtered_rows``,
``_dsv4_normalize_dtype``, the DSA model-dim constants) now live next
to the op classes that own them, completing Pattern A's "data + load +
query lives together" guarantee that the rest of the refactor's design
already assumed.

Per-module distribution after the move:

- ``gemm.py`` gains 3 loaders (gemm / compute_scale / scale_matrix)
- ``attention.py`` gains 2 (context / generation)
- ``moe.py`` gains 7 (regular + 6 WideEP / TRT-LLM variants)
- ``mla.py`` gains 7 (regular + module + bmm + WideEP variants)
- ``dsa.py`` gains 2 + the ``DSA_MODEL_DIMS`` / ``DEFAULT_DSA_ARCHITECTURE``
  constants
- ``dsv4.py`` gains 4 loaders + ``_dsv4_normalize_dtype`` + ``_DSV4_DTYPE_ALIASES``
- ``mamba.py`` gains 2 (mamba2 / gdn)
- ``communication.py`` gains 2 (custom_allreduce / nccl)

Supporting changes:

- ``_read_filtered_rows`` moves to ``operations/base.py``. Every op
  module's loader uses it, and putting it on the operations side
  eliminates a circular ``operations/X → perf_database → operations/X``
  module-load cycle that would otherwise fire when an op module is the
  import entry point.
- The op-class ``load_data`` methods drop their function-local
  ``from ...perf_database import load_X_data`` lines: the loaders are
  now module-local and resolve via the ordinary global lookup, so
  ``mock.patch("aiconfigurator.sdk.operations.<mod>.load_X_data", ...)``
  actually intercepts the call. ``LoadedOpData`` and
  ``PerfDataFilename`` stay in ``perf_database`` and continue to be
  imported function-locally.
- ``LazySupportMatrix`` already lives in ``perf_database.py`` from the
  previous commit; resolvers continue to call op-class ``load_data``
  via the package, no change needed.

Test surface:

- ``conftest._LOADER_STUBS`` becomes ``{name: (target_module, default_return)}``.
  Both ``_patch_all_loaders_and_yaml()`` and
  ``_get_comprehensive_db_singleton()`` patch the new locations
  (``aiconfigurator.sdk.operations.<mod>.load_X_data``).
- Three test files with explicit ``monkeypatch.setattr(
  "aiconfigurator.sdk.perf_database.load_X_data", ...)`` sites
  (``test_comm_fallback.py``, ``test_custom_allreduce_empty_bucket.py``,
  ``test_edge_cases.py``) update to the new paths.
- ``test_clear_all_op_caches_resets_counter_and_class_cache`` snapshots
  and restores EVERY op subclass's class-level caches (not just GEMM's)
  because per-op-class ``_query_*_table`` now calls
  ``cls.load_data(database)`` at the top, so wiping any op's class
  cache breaks the ``comprehensive_perf_db`` singleton for later tests.

Part of AIC-533 (Phase 4.5 cleanup, follow-on commit).
