+++ src/aiconfigurator/sdk/operations/dsa.py
@@ -29,16 +29,45 @@ from __future__ import annotations
 
 import copy
 import logging
+from collections import defaultdict
 from typing import TYPE_CHECKING, ClassVar
 
 from aiconfigurator.sdk import common, interpolation
-from aiconfigurator.sdk.operations.base import Operation
+from aiconfigurator.sdk.operations.base import Operation, _read_filtered_rows
 from aiconfigurator.sdk.performance_result import PerformanceResult
 
 if TYPE_CHECKING:
     from aiconfigurator.sdk.perf_database import PerfDatabase
 
 logger = logging.getLogger(__name__)
+
+
+DSA_MODEL_DIMS: dict[str, dict] = {
+    "DeepseekV32ForCausalLM": {
+        "hidden_size": 7168,
+        "q_lora_rank": 1536,
+        "kv_lora_rank": 512,
+        "qk_nope_head_dim": 128,
+        "qk_rope_head_dim": 64,
+        "v_head_dim": 128,
+        "index_topk": 2048,
+        "index_head_dim": 128,
+        "index_n_heads": 64,
+    },
+    "GlmMoeDsaForCausalLM": {
+        "hidden_size": 6144,
+        "q_lora_rank": 2048,
+        "kv_lora_rank": 512,
+        "qk_nope_head_dim": 192,
+        "qk_rope_head_dim": 64,
+        "v_head_dim": 256,
+        "index_topk": 2048,
+        "index_head_dim": 128,
+        "index_n_heads": 32,
+    },
+}
+
+DEFAULT_DSA_ARCHITECTURE = "DeepseekV32ForCausalLM"
 
 
 # Extrapolation grids — lifted verbatim from the legacy blocks in
@@ -172,7 +201,7 @@ class ContextDSAModule(Operation):
         ``database._raw_context_dsa_module_data``."""
         import os
 
-        from aiconfigurator.sdk.perf_database import LoadedOpData, PerfDataFilename, load_context_dsa_module_data
+        from aiconfigurator.sdk.perf_database import LoadedOpData, PerfDataFilename
 
         key = cls._cache_key(database)
         if key not in cls._data_cache:
@@ -524,7 +553,7 @@ class GenerationDSAModule(Operation):
         extrapolation, binds ``database._generation_dsa_module_data``."""
         import os
 
-        from aiconfigurator.sdk.perf_database import LoadedOpData, PerfDataFilename, load_generation_dsa_module_data
+        from aiconfigurator.sdk.perf_database import LoadedOpData, PerfDataFilename
 
         key = cls._cache_key(database)
         if key not in cls._data_cache:
@@ -761,3 +790,104 @@ class GenerationDSAModule(Operation):
 
     def get_weights(self, **kwargs):
         return self._weights * self._scale_factor
+
+
+# ─────────────────────────────────────────────────────────
+# CSV loaders (moved from perf_database.py in AIC-533 cleanup)
+# ─────────────────────────────────────────────────────────
+
+
+def load_context_dsa_module_data(dsa_file: str):
+    """
+    Load context DSA data.
+
+    Dict structure:
+        data[fmha_quant_mode][kv_cache_quant_mode][gemm_quant_mode][architecture][num_heads][s][b]
+
+    Quant modes are the outermost keys so that ``_enum_key_names`` can
+    directly extract supported FMHAQuantMode names (aligned with
+    ``_context_attention_data``).  ``architecture`` (e.g.
+    "DeepseekV32ForCausalLM", "GlmMoeDsaForCausalLM") selects the
+    model-specific structural dimensions from ``DSA_MODEL_DIMS``.
+    Legacy CSV rows without an ``architecture`` column default to
+    "DeepseekV32ForCausalLM".
+    """
+    rows = _read_filtered_rows(dsa_file)
+    if rows is None:
+        logger.debug(f"DSA context data file {dsa_file} not found.")
+        return None
+
+    dsa_data = defaultdict(
+        lambda: defaultdict(
+            lambda: defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: defaultdict()))))
+        )
+    )
+
+    has_power = len(rows) > 0 and "power" in rows[0]
+
+    for row in rows:
+        num_heads = int(row["num_heads"])
+        b = int(row["batch_size"])
+        s = int(row["isl"])
+        latency = float(row["latency"])
+        power = float(row.get("power", 0.0)) if has_power else 0.0
+        energy = power * latency
+
+        arch = row.get("architecture", DEFAULT_DSA_ARCHITECTURE)
+        gemm_mode = common.GEMMQuantMode[row["gemm_type"]]
+        fmha_mode = common.FMHAQuantMode[row["mla_dtype"]]
+        kv_dtype = common.KVCacheQuantMode[row["kv_cache_dtype"]]
+
+        dsa_data[fmha_mode][kv_dtype][gemm_mode][arch][num_heads][s][b] = {
+            "latency": latency,
+            "power": power,
+            "energy": energy,
+        }
+
+    return dsa_data
+
+
+def load_generation_dsa_module_data(dsa_file: str):
+    """
+    Load generation DSA data.
+
+    Dict structure:
+        data[kv_cache_quant_mode][gemm_quant_mode][architecture][num_heads][b][s]
+
+    Quant modes are the outermost keys so that ``_enum_key_names`` can
+    directly extract supported KVCacheQuantMode names (aligned with
+    ``_generation_attention_data``).  ``architecture`` selects the
+    model-specific structural dimensions from ``DSA_MODEL_DIMS``.
+    Legacy CSV rows without an ``architecture`` column default to
+    "DeepseekV32ForCausalLM".
+    """
+    rows = _read_filtered_rows(dsa_file)
+    if rows is None:
+        logger.debug(f"DSA generation data file {dsa_file} not found.")
+        return None
+
+    dsa_data = defaultdict(
+        lambda: defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: defaultdict()))))
+    )
+
+    has_power = len(rows) > 0 and "power" in rows[0]
+
+    for row in rows:
+        num_heads = int(row["num_heads"])
+        b = int(row["batch_size"])
+        s = int(row["isl"]) + int(row["step"])
+        latency = float(row["latency"])
+        power = float(row.get("power", 0.0)) if has_power else 0.0
+        energy = power * latency
+
+        arch = row.get("architecture", DEFAULT_DSA_ARCHITECTURE)
+        gemm_mode = common.GEMMQuantMode[row["gemm_type"]]
+        kv_dtype = common.KVCacheQuantMode[row["kv_cache_dtype"]]
+
+        dsa_data[kv_dtype][gemm_mode][arch][num_heads][b][s] = {
+            "latency": latency,
+            "power": power,
+            "energy": energy,
+        }
+
+    return dsa_data
