+++ tests/unit/sdk/database/test_comm_fallback.py
@@ -102,19 +102,19 @@ def _db_factory(tmp_path, monkeypatch):
                 return nccl_data
             return oneccl_data
 
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_nccl_data", _nccl_load_dispatch)
+        monkeypatch.setattr("aiconfigurator.sdk.operations.communication.load_nccl_data", _nccl_load_dispatch)
 
         # Patch all other loaders to avoid file access
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_gemm_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_context_attention_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_generation_attention_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_moe_data", lambda p: ({}, {}))
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_custom_allreduce_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_context_mla_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_generation_mla_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_mla_bmm_data", lambda p: {})
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_context_dsa_module_data", lambda p: None)
-        monkeypatch.setattr("aiconfigurator.sdk.perf_database.load_generation_dsa_module_data", lambda p: None)
+        monkeypatch.setattr("aiconfigurator.sdk.operations.gemm.load_gemm_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.attention.load_context_attention_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.attention.load_generation_attention_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.moe.load_moe_data", lambda p: ({}, {}))
+        monkeypatch.setattr("aiconfigurator.sdk.operations.communication.load_custom_allreduce_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.mla.load_context_mla_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.mla.load_generation_mla_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.mla.load_mla_bmm_data", lambda p: {})
+        monkeypatch.setattr("aiconfigurator.sdk.operations.dsa.load_context_dsa_module_data", lambda p: None)
+        monkeypatch.setattr("aiconfigurator.sdk.operations.dsa.load_generation_dsa_module_data", lambda p: None)
 
         yaml_file = tmp_path / "sys.yaml"
         yaml_file.write_text("dummy: data")
